+++ src/admin_commands/ban.js
@@ -45,17 +45,18 @@ module.exports = {
             await banMessage.react(green_check);
             await banMessage.react(red_x);
             const collector = banMessage.createReactionCollector(filter, { maxEmojis: 2 });
-            collector.on('collect', (reaction, user) => {
+            collector.on('collect', async (reaction, user) => {
                 switch (reaction.emoji.name) {
                     case green_check:
                         if (user.id === constants.botId) break;
                         memberTarget.ban();
-                        banfunction.onKickBan(firedb, memberTarget.user.tag, message.channel, id);
+                        await messages.server_log(firedb, client, message.guild.id, `<@${id}> has been banned by <@${message.author.id}`);
+                        await banfunction.onKickBan(firedb, memberTarget.user.tag, message.channel, id);
                         collector.stop();
                         break;
                     default:
                         if (user.id === constants.botId) break;
-                        messages.send_message(firedb, message.channel, `<@${id}> lives to see another day`);
+                        await messages.send_message(firedb, message.channel, `<@${id}> lives to see another day`);
                         collector.stop();
                         break;
                 }