+++ src/events/guild/channelUpdate.js
@@ -1,7 +1,7 @@
 const databasefunctions = require('../../functions/databasefunctions.js');
-const eventlogging = require('../../functions/eventlogging.js');
+const messages = require('../../functions/messages.js');
 
 module.exports = async (Discord, client, firedb, oldChannel, newChannel) => {
     await databasefunctions.UpdateChannelInServer(firedb, oldChannel.guildId, oldChannel.id, newChannel.id, newChannel.name);
-    await eventlogging.ServerLog(firedb, client, oldChannel.guildId, '**' + oldChannel.name + '** channel has been updated');
+    await messages.server_log(firedb, client, oldChannel.guildId, '**' + oldChannel.name + '** channel has been updated');
 };